requirements.txt
gradio
easyocr
opencv-python-headless
numpy
Pillow
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu